get demo working through the debug adapter
demo is now fetching and loading the example story code and can logs
the initial story state in the console.

removed initial state property and made it work more like react where
you set the state in the constructor. This is all levels of the
inheritance hierarchy can initialize their own state. This is great for
things like mixins that will add their functionality to items.