moved some of demo's index into an app component
This way the App component can maintain the story state, the additional
app state, render chrome, and react to adapter callbacks.

This lets the display be completely controlled by sending it the whole
of the story state and it won't have to deal with the adapter at all,
which is really an application concern since the app creates the
adapter.