reduced bundle size by cleaning up dependencies
React and some other large packages were being included in the bundle
twice or more. One reason for this is because of the symlinks provided
by lerna changes the normal package lookup: symlinked packages wind up
finding packages in the root level node_modules instead of in the
node_modules of the package importing it. Finding packages in distinct
paths looks like two different packages to webpack.

Also fixed some of the externals so some of this packages lookup stuff
doesn't need to happen.