remove unnecessary library building
having webpacked files inside webpacked files is kinda wonky, especially
when it comes to source maps and finding original files to debug in the
browser. Only building the libraries as part of the final webpack makes
this much easier, and the babel configuration was shared anyway.

Anything outside of the watif repo that wants to use these packages will
have to build the src with appropriate babel rules. Maybe later both the
compiled and uncompiled versions can be provided.